+++ client/app/pods/components/nested-question-textarea/component.js
@@ -6,6 +6,13 @@ export default NestedQuestionComponent.extend({
   unwrappedHelpText: null,
   displayContent: true,
   inputClassNames: ['form-control'],
+  browserDetector: Ember.inject.service(),
+
+  input() {
+    if (this.get('browserDetector.isIE11OrLess')) {
+      this.save();
+    }
+  },
 
   clearHiddenQuestions: Ember.observer('displayContent', 'disabled', function() {
     if (!this.get('disabled') && !this.get('displayContent')) {