+++ spec/factories/attachment_factory.rb
@@ -14,6 +14,23 @@ FactoryGirl.define do
       attachment.owner ||= FactoryGirl.create(:ad_hoc_task)
     end
 
+    trait :processing do
+      status "processing"
+    end
+
+    trait :errored do
+      status "error"
+      error_message "Failed for some reason"
+    end
+
+    trait :completed do
+      status "done"
+    end
+
+    trait :unknown_state do
+      status "unknown"
+    end
+
     trait :with_resource_token do
       after :build do |attachment|
         attachment.build_resource_token(attachment.file)